feat(engine): move task headers to user task properties
The task headers extension elements need to be transported in the general user task properties.
Job worker properties will inerit this from there and keep on working as before.
The user task processor will be able to fetch the header to store them in user task records.
